moving total balance files to specific package
